Creating company,branch and posts controller in admin panel

+++ src/Controller/PostsController.php
@@ -16,6 +16,7 @@ class PostsController extends AppController
      *
      * @return \Cake\Network\Response|null
      */
+    //The default page of post load by this function.
     public function index()
     {
         $post = $this->paginate($this->Posts);
@@ -30,11 +31,20 @@ class PostsController extends AppController
      * @return \Cake\Network\Response|null
      * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
      */
+     //User can view the data of perticular post.
     public function view($id = null)
     { 
         $post = $this->Posts->get($id, [
             'contain' => []
         ]);
+
+        //  $user = $this->request->session()->read('Auth.User.username');
+        // $post=$comments->find('all')
+        //          ->where(['Posts.post_id' => $id ])
+        //          ->contain('Users')
+        //          ->toArray();
+        //          $this->set(compact('post'));
+
         $this->set('post', $post);
         $this->set('_serialize', ['post']);
     }
@@ -44,14 +54,17 @@ class PostsController extends AppController
      *
      * @return \Cake\Network\Response|null Redirects on successful add, renders view otherwise.
      */
+     //User can add the post by this function.
     public function add()
     {
+        // $authuser_id = $this->request->session()->read('Auth.User.id');
         $post = $this->Posts->newEntity();
         if ($this->request->is('post')) {
 
             $post = $this->Posts->patchEntity($post, $this->request->getData());
             $post->created = date("Y-m-d H:i:s");
             $post->modified = date("Y-m-d H:i:s");
+            $post->user_id = $this->request->session()->read('Auth.User.id');
             if ($this->Posts->save($post)) {
                 $this->Flash->success(__('The post has been saved.'));
 
@@ -70,6 +83,7 @@ class PostsController extends AppController
      * @return \Cake\Network\Response|null Redirects on successful edit, renders view otherwise.
      * @throws \Cake\Network\Exception\NotFoundException When record not found.
      */
+     //User can edit the post data by edit function.
     public function edit($id = null)
     {
         $post = $this->Posts->get($id, [
@@ -95,20 +109,31 @@ class PostsController extends AppController
      * @return \Cake\Network\Response|null Redirects to index.
      * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
      */
+    //User can delete the perticular post by delete function.
     public function delete($id = null)
     {
-        $this->request->allowMethod(['post', 'delete']);
+        $userid = $this->request->session()->read('Auth.User.id');
         $post = $this->Posts->get($id);
-        if ($this->Posts->delete($post)) {
-            $this->Flash->success(__('The post has been deleted.'));
-        } else {
-            $this->Flash->error(__('The post could not be deleted. Please, try again.'));
-        }
+        if($userid == $post['user_id']){
+            // print_r($userid);exit;
+            $this->request->allowMethod(['post', 'delete']);
+            $post = $this->Posts->get($id);
+            if ($this->Posts->delete($post)) {
+                $this->Flash->success(__('The post has been deleted.'));
+            } else {
+                $this->Flash->error(__('The post could not be deleted. Please, try again.'));
+            }
 
-        return $this->redirect(['action' => 'index']);
+            return $this->redirect(['action' => 'index']);
+        }
+        else{
+            $this->Flash->error(__('You are not authorized.'));
+            return $this->redirect(['action' => 'index']);
+        }
     }
 
-     public function comment($id = null)
+    //User can perform comment operation through this function.
+     public function comment($id ,$comment_id= null)
     {
         $post = $this->Posts->get($id, [
             'contain' => []
